% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_data.R
\docType{data}
\name{aggregationData}
\alias{aggregationData}
\title{Aggregation Data}
\format{An object of class \code{matrix} with 788 rows and 2 columns.}
\source{
\href{https://cs.joensuu.fi/sipu/datasets/}{Clustering basic benchmark}
}
\usage{
data(aggregationData, package = "jDirichletMixtureModels")
}
\description{
N=788, k=7, D=2
}
\examples{
\dontrun{
Xdata <- data(aggregationData, package = "jDirichletMixtureModels")
states <- dmm.cluster(model, Xdata)
dmm.plot(states[[1]]$labeledData)
}
}
\references{
A. Gionis, H. Mannila, and P. Tsaparas, Clustering aggregation. ACM Transactions on Knowledge Discovery from Data (TKDD), 2007. 1(1): p. 1-30.
}
\keyword{datasets}
