% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_data.R
\docType{data}
\name{birch3Data}
\alias{birch3Data}
\title{Birch3 Data}
\format{An object of class \code{matrix} with 100000 rows and 2 columns.}
\source{
\href{https://cs.joensuu.fi/sipu/datasets/}{Clustering basic benchmark}
}
\usage{
data(birch3Data, package = "jDirichletMixtureModels")
}
\description{
Synthetic 2-d data with N=100,000 vectors and k=100 clusters. Random sized clusters in random locations.
}
\examples{
\dontrun{
Xdata <- data(birch3Data, package = "jDirichletMixtureModels")
states <- dmm.cluster(model, Xdata)
dmm.plot(states[[1]]$labeledData)
}
}
\references{
Zhang et al., "BIRCH: A new data clustering algorithm and its applications", Data Mining and Knowledge Discovery, 1 (2), 141-182, 1997.
}
\keyword{datasets}
