% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_model.R
\name{dmm.addfile}
\alias{dmm.addfile}
\title{Add a Julia file to be accessible.}
\usage{
dmm.addfile(filename)
}
\arguments{
\item{filename}{The name of the Julia file.}
}
\description{
This function must be called before runnning the \code{dmm.cluster} function using a JModel (see \code{dmm.JModel}).
}
