% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_model.R
\name{dmm.BaseModel}
\alias{dmm.BaseModel}
\title{Create a model using bulit-in conjugate models}
\usage{
model <- dmm.BaseModel(typename, params)
}
\arguments{
\item{typename}{A string. The name of the predefined conjugate prior you wish to use. Options listed under details.
"MultivariateNormalModel" is the default.}

\item{params}{A list of the hyperparameter values for the likelihood functions. Many model have default \code{params} values and thus can be made without passing any \code{params}.
See documentation for what parameters a given model may take.}

\item{data}{In lieu of explicit hyperparameter values, some models can infer good hyperparameter values from given data.
This option is supported for MultivariateNormalModel and UnivariateNormalModel, as well as UnivariateNormalKnownSigma.}
}
\value{
A model object of type BaseModel which can be passed to \code{dmm.cluster}.
}
\description{
Create an model object to be used in the \code{dmm.cluster} function, using the packages bulit-in conjugate models.
Function \code{dmm.model} is an alternative method.
}
\details{
Bulit-in models avaible are:
                       "MultivariateNormalModel" (default),
                       "UnivariateNormalModel",
                       "UnivariateNormalKnownSigma",
                       "UnivariateExponentialModel".
}
