% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{dmm.cluster.BaseModel}
\alias{dmm.cluster.BaseModel}
\title{Use a Dirichlet Mixture Model on data to get cluster labels and cluster parameter values.}
\usage{
\method{dmm.cluster}{BaseModel}(model, Xdata, alpha = 1, iters = 5000,
  burnin = 200, shuffle = TRUE)
}
\description{
If using one of the avaible bulit-in models.
}
