% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_model.R
\name{dmm.JModel}
\alias{dmm.JModel}
\title{Create a model using Julia fucntions}
\usage{
\code{dmm.addfile(filename)}
\code{model <- dmm.model(pdf_name, sample_name, marg_name, params, isconjugate=TRUE)}
\code{model <- dmm.Jmodel(pdf_name, sample_name, marg_name, params, isconjugate=TRUE)}
\code{model <- dmm.JConjugateModel(pdf_name, sample_name, marg_name, params)}
\code{model <- dmm.JNonConjugateModel(pdf_name, sample_name, params)}
}
\arguments{
\item{pdf_name}{A string. The name of the Julia function in \code{filename} that returns the probability density function likelihood. The function should be of the form \code{pdf_name(y::Float64, theta::Tuple, params::Tuple)} or \code{pdf_name(y::Array{Float64,1}, theta::Tuple, params::Tuple)}.}

\item{sample_name}{A string. The name of the Julia function in \code{filename} that returns the sample posterior function for conjugate case or the sample prior for nonconjugate case. The function should be of the form \code{sample_name(y::Float64, params::Tuple)}, \code{sample_name(y::Array{Float64,1}, params::Tuple)} or \code{sample_name(y::Array{Float64,2}, params::Tuple)}.}

\item{marg_name}{A string. For conjugate case only. The name of the Julia function in \code{filename} that returns the marginal likelihood. The function should be of the form \code{marg_name(y::Float64, params::Tuple)}.}

\item{params}{A list of all hyperparameters needed for the above three functions.}

\item{isconjugate}{A logical. \code{TRUE} (default) if the user specfied model is conjugate, \code{FALSE} if not.}
}
\value{
A model object of type JModel which can be passed to \code{dmm.cluster}.
}
\description{
Create an model object to be used in the \code{dmm.cluster} function, using user given Julia functions. Must call \code{dmm.addfile} to import files, in which the Julia functions are stored, before using \code{dmm.cluster} on a JModel.
Functions \code{dmm.JConjugateModel} and \code{dmm.JNonConjugateModel} are alternatives.
}
\details{
\code{marg_name} is only requried for conjugate models.
}
\examples{
dmm.addfile(filename)
# The following all make models using Julia functions
model <- dmm.model(pdf_name, sample_name, marg_name, params, isconjugate=TRUE)
model <- dmm.Jmodel(pdf_name, sample_name, marg_name, params, isconjugate=TRUE)
model <- dmm.JConjugateModel(pdf_name, sample_name, marg_name, params)
model <- dmm.JNonConjugateModel(pdf_name, sample_name, params)

}
