% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_model.R
\name{dmm.model.character}
\alias{dmm.model.character}
\title{Create a model object}
\usage{
\method{dmm.model}{character}(...)
}
\description{
If a string is passed as the first input to \code{dmm.model} it will create and return either a BaseModel (if passed one string, plus an optional \code{params}) (see \code{dmm.model.BaseModel}),
or a JModel (see \code{dmm.model.JModel})
}
