% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{dmm.model.function}
\alias{dmm.model.function}
\title{Create a model object to be used for dmm.cluster
If a function is passed to dmm.model it will create and return an Rmodel}
\usage{
\method{dmm.model}{function}(...)
}
\description{
Create a model object to be used for dmm.cluster
If a function is passed to dmm.model it will create and return an Rmodel
}
