% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_model.R
\name{dmm.model.NULL}
\alias{dmm.model.NULL}
\title{Create a model object}
\usage{
\method{dmm.model}{NULL}(...)
}
\description{
If nothing is passed to \code{dmm.model} it will create and return a BaseModel using a multivariate normal conjugate model with default parameters (see \code{dmm.BaseModel}).
}
