% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_model.R
\name{dmm.model}
\alias{dmm.model}
\title{Create a model}
\usage{
\code{dmm.model(...)}
}
\value{
A model object which can be passed to \code{dmm.cluster}.
}
\description{
Make a model object to be passed to \code{dmm.cluster}. If not passed any argurments a conjugate multivariate normal likelihood model with default parameters will be used.
}
\details{
Depending on what argurments are passed to \code{dmm.model()} either a BaseModel (see \code{dmm.BaseModel}),
a RModel (see \code{dmm.RModel}), or JModel (see \code{dmm.JModel}) will be made.
}
\examples{
\dontrun{
# Example of using a multivariate normal conjugate prior (the default)
model <- dmm.model()

# Example of using a bulit-in conjugate prior
model <- dmm.model(typename, params)

# Using a user specified R functions
model <- dmm.model(pdf_fct, sample_fct, marg_fct, params, isconjugate=TRUE)

# Using a user specified Julia functions located in a .jl file called filename
dmm.addfile(filename)
model <- dmm.model(pdf_name, sample_name, marg_name, params, isconjugate=TRUE)
}

}
