% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_display.R
\name{dmm.plot}
\alias{dmm.plot}
\title{Plot labeledData for a state returned by dmm}
\usage{
\code{ states <- dmm.cluster(model,Xdata,...)  }
\code{ dmm.plot(states[[1]]$labeledData)  }

Given the labeledData from a single state returned by dmm.cluster(...), plot it. Can do 2D, 1D, or 3D plots.
\code{ggplot2} recommanded for 2D plots. \code{scatterplot3d} requried for 3D plots.
}
\arguments{
\item{labeledData}{The labeledData from a single state. A single state is one item of the list of states deturned by mm.cluster(...).}
}
\description{
Plot labeledData for a state returned by dmm
}
