% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_model.R
\name{dmm.RModel}
\alias{dmm.RModel}
\title{Create a model using R fucntions}
\usage{
dmm.RModel(pdf_func, sample_func = NULL, marg_func = NULL, params,
  isconjugate = TRUE)
}
\arguments{
\item{pdf_func}{A function that takes as input \code{(data, likelihoodparams, params)}. It returns the value of the probability density function likelihood at \code{(data,likelihoodparams)} given \code{params}.}

\item{params}{A list of all hyperparameters needed for the above three functions.}

\item{isconjugate}{A logical. \code{TRUE} (default) if the user specfied model is conjugate, \code{FALSE} if not.}

\item{sample_fct}{Optional for nonconjugate models. A function takes as input \code{(data, params)}. It returns the value of the sample posterior function at \code{data} given \code{params}.}

\item{marg_fct}{Only needed for conjugate models. A function takes as input \code{(data, params)}. It returns the value of the marginal likelihood funciton at \code{data} given \code{params}..}
}
\value{
A model object of type RModel which can be passed to \code{dmm.cluster}.
}
\description{
Create an model object to be used in the \code{dmm.cluster} function, using user given R functions.
}
\details{
@usage \code{model <- dmm.Rmodel(pdf_fct, sample_fct, marg_fct, params, isconjugate=TRUE)}
}
