% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_setup.R
\name{dmm.setup}
\alias{dmm.setup}
\title{Do initial setup for jDirichletMixtureModels package.}
\usage{
dmm.setup()
}
\arguments{
\item{...}{arguments passed to \code{JuliaCall::julia_setup}. Most noteable:the first input should be the path to the folder in which the Julia bin file is located on the user's machine. 
It will default to using the macOSX path for the most recent Julia version.}
}
\description{
\code{dmm.setup} does the initial setup for jDirichletMixtureModels package.
}
\examples{
\dontrun{
dmm.setup()
}

}
