% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_cluster.R
\name{dmm.state}
\alias{dmm.state}
\title{Formats a single DMM OutputState Julia object into a list of labeled data and list of cluster info}
\usage{
dmm.state(juliastate, paramnames)
}
\description{
Formats a single DMM OutputState Julia object into a list of labeled data and list of cluster info
}
