% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{dmm.stateAsTable}
\alias{dmm.stateAsTable}
\title{Formats a single DMM OutputState Julia object into a list of labeled data and data.table of cluster info}
\usage{
dmm.stateAsTable(juliastate, paramnames = NULL)
}
\description{
Formats a single DMM OutputState Julia object into a list of labeled data and data.table of cluster info
}
