% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{dmm.states}
\alias{dmm.states}
\title{Constructing list of all states from dmm.cluster run (excluding burnin)}
\usage{
dmm.states(juliastates, paramnames = NULL)
}
\description{
Constructing list of all states from dmm.cluster run (excluding burnin)
}
