% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_display.R
\name{dmm.summarize.list}
\alias{dmm.summarize.list}
\title{Summerize given a single state's cluster info}
\usage{
dmm.summarize(clusterInfo)
}
\arguments{
\item{clusterInfo}{A list. Given states <- dmm.cluster(...), this input is states[[i]]$clusters}
}
\description{
Summerize given a single state's cluster info
}
\examples{
states <- dmm.cluster(model, data)
dmm.summarize(states[[1]]$clusters)

}
