% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_display.R
\name{dmm.summarize}
\alias{dmm.summarize}
\title{Summerize given a single state's cluster info}
\usage{
dmm.summarize(clusterInfo)
}
\arguments{
\item{clusterInfo}{A list or data.table. Given states <- dmm.cluster(...), this input is states$clusterInfo}
}
\description{
Summerize given a single state's cluster info
}
