% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{.add_file}
\alias{.add_file}
\title{Including a user given file of Julia functions}
\usage{
.add_file(filename)
}
\arguments{
\item{filename}{The name of the Julia file}
}
\description{
Including a user given file of Julia functions
}
