% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{.BaseModel}
\alias{.BaseModel}
\title{Constructing BaseModel object: a conjugate model given by typename}
\usage{
.BaseModel(typename = "MultivariateNormal", params = NULL)
}
\description{
Constructing BaseModel object: a conjugate model given by typename
}
