% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{.JModel}
\alias{.JModel}
\title{Constructing JModel object: a model based on user specifed Julia functions}
\usage{
.JModel(pdf_name, sample_name, marg_name, params, isconjugate = TRUE)
}
\description{
Constructing JModel object: a model based on user specifed Julia functions
}
