% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{.RModel}
\alias{.RModel}
\title{Constructing RModel object: a model based on user specifed R function}
\usage{
.RModel(pdf_func, sample_func, marg_func, params, isconjugate = TRUE)
}
\description{
Constructing RModel object: a model based on user specifed R function
}
