% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_data.R
\docType{data}
\name{syntheticGaussian2}
\alias{syntheticGaussian2}
\title{Synthetic Gaussian Cluster Data 2}
\format{An object of class \code{matrix} with 5000 rows and 2 columns.}
\source{
\href{https://cs.joensuu.fi/sipu/datasets/}{Clustering basic benchmark}
}
\usage{
data("syntheticGaussian2", package = "jDirichletMixtureModels")
}
\description{
Synthetic 2-d data with N=5000 vectors and k=15 Gaussian clusters with different degree of cluster overlapping
}
\examples{
\dontrun{
Xdata <- data(syntheticGaussian2)
states <- dmm.cluster(model, Xdata)
dmm.plot(states[[1]]$labeledData)
}
}
\references{
P. Fränti and O. Virmajoki, "Iterative shrinking method for clustering problems", Pattern Recognition, 39 (5), 761-765, May 2006.
}
\keyword{datasets}
