% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_data.R
\docType{data}
\name{unbalancedClusters}
\alias{unbalancedClusters}
\title{Synthetic Unbalanced Gaussian Cluster Data}
\format{An object of class \code{matrix} with 6500 rows and 2 columns.}
\source{
\href{https://cs.joensuu.fi/sipu/datasets/}{Clustering basic benchmark}
}
\usage{
data(unbalancedClusters)
}
\description{
Synthetic 2-d data with N=6500 vectors and k=8 Gaussian clusters
}
\examples{
\dontrun{
Xdata <- data(unbalancedClusters)
states <- dmm.cluster(model, Xdata)
dmm.plot(states[[1]]$labeledData)
}
}
\references{
M. Rezaei and P. Fränti, "Set-matching methods for external cluster validity", IEEE Trans. on Knowledge and Data Engineering, 28 (8), 2173-2186, August 2016.
}
\keyword{datasets}
