% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{dmm.cluster}
\alias{dmm.cluster}
\title{Use a Dirichlet Mixture Model on data to get cluster labels and cluster parameter values.}
\usage{
dmm.cluster(model, data, alpha = 1, iters = 5000, burnin = 200,
  shuffle = TRUE)
}
\arguments{
\item{model}{An object returned by \code{dmm.model()}.}

\item{data}{A 1D array of length N (univariate case) or 2D array of size d-by-N (mulitvariate case),
where d is the dimensionailty of the data and N is the number of observations.}

\item{alpha}{}

\item{iters}{An integer. Number of iterations.}

\item{burnin}{An integer. Amount of burn-in.}

\item{shuffle}{A logical. Whether or not to shuffle the data.}
}
\description{
Use a Dirichlet Mixture Model on data to get cluster labels and cluster parameter values.
}
