% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{dmm.cluster.RModel}
\alias{dmm.cluster.RModel}
\title{Use a Dirichlet Mixture Model on data to get cluster labels and cluster parameter values.
If using a user specifed model via R functions.}
\usage{
\method{dmm.cluster}{RModel}(model, Xdata, alpha = 1, m_prior = 3,
  m_post = 3, iters = 5000, burnin = 200, shuffle = TRUE)
}
\description{
Use a Dirichlet Mixture Model on data to get cluster labels and cluster parameter values.
If using a user specifed model via R functions.
}
