% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{dmm.model.character}
\alias{dmm.model.character}
\title{Create a model object to be used for dmm.cluster
If a string is passed to dmm.model it will create and return either a BaseModel (if its one string),
or a JModel for 4 strings}
\usage{
\method{dmm.model}{character}(...)
}
\description{
Create a model object to be used for dmm.cluster
If a string is passed to dmm.model it will create and return either a BaseModel (if its one string),
or a JModel for 4 strings
}
