% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_model.R
\name{dmm.model.function}
\alias{dmm.model.function}
\title{Create a model object}
\usage{
\method{dmm.model}{function}(...)
}
\description{
If a function is passed as the first input to \code{dmm.model} it will create and return an Rmodel (see \code{dmm.RModel}).
}
