% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{dmm.model}
\alias{dmm.model}
\title{Create the model used in the dmm.cluster function}
\usage{
dmm.model(...)
}
\arguments{
\item{typename}{A string. The name of the predefined conjugate prior you wish to use. Options are:
        "MultivariateNormal" (default) 
        "UnivariateNormal"
        "UnivariateNormalKnowSigma" - requires \code{params} = sigma
        "UnivariateExponential"
These do not require any \code{params} unless otherwise stated}

\item{pdf_func}{A function. The probability density function likelihood}

\item{sample_fct}{A function. The sample posterior function}

\item{marg_fct}{A function. The marginal likelihood}

\item{filename}{A string. The name of the .jl file in which Julia functions are located}

\item{pdf_name}{A string. The name of the Julia function in \code{filename} that returns the probability density function likelihood.}

\item{sample_name}{A string. The name of the Julia function in \code{filename} that returns the sample posterior function.}

\item{marg_name}{A string. The name of the Julia function in \code{filename} that returns the marginal likelihood.}

\item{params}{A list of the hyperparameter values used by \code{pdf_func}, \code{sample_func}, and \code{marg_func}.}

\item{uv}{A logical variable. \code{TRUE} (default) if model is univariate,\code{FALSE} if not.}

\item{isconjugate}{A logical variable.\code{TRUE} if the user specfied model is conjugate, \code{FALSE} (default) if not.}
}
\value{
A model object which can be passed to \code{dmm.cluster}.
}
\description{
If not passed any argurments a multivariate normal conjugate prior will be used as default.
}
\details{
Either just the name of one of the availble conjugate priors should be passed as \code{typename} (along with any hyperparamters if needed),
or the user can specify their own model. 
User specified models require three R functions: \code{pdf_fct}, \code{sample_fct}, and \code{marg_fct}
(and hyperparameters needed for each of these in \code{params}). 
Users can specify these functions as Julia functions if desired. To do this pass the filename of the .jl file in which
the functions are saved, along with the names of the three functions in the given Julia code.
}
\examples{
\dontrun{
# Example of using a multivariate normal conjugate prior (the default)
model <- dmm.model()

# Example of using a bulit-in conjugate prior
model <- dmm.model(typename = "UnivariateNormalKnowSigma", params = list(sigma))

# Using a user specified R functions 
model <- dmm.model(pdf_fct, sample_fct, marg_fct, params, uv=TRUE, isconjugate=TRUE)
                                
# Using a user specified Julia functions 
model <- dmm.model(filename, pdf_name, sample_name, marg_name, params, uv=TRUE, isconjugate=TRUE)
}  

}
