% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_display.R
\name{dmm.plot}
\alias{dmm.plot}
\title{Plot labeledData for a state returned by dmm}
\usage{
dmm.plot(labeledData)  

Given the data from a single state returned by dmm.cluster(...), plot it. Can do 2D, 1D, or 3D plots.
\code{ggplot2} recommanded for 2D plots. \code{scatterplot3d} requried for 3D plots.
}
\arguments{
\item{labeledData}{The data from a single state. Given states <- dmm.cluster(...), this input is states[[i]]$data.}
}
\description{
Plot labeledData for a state returned by dmm
}
\examples{
states <- dmm.cluster(model,Xdata,...)  
 dmm.plot(states[[1]]$data)  

}
