% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_functions.R
\name{dmm.setup}
\alias{dmm.setup}
\title{Do initial setup for jDirichletMixtureModels package.}
\usage{
dmm.setup(...)
}
\arguments{
\item{...}{arguments passed to \code{JuliaCall::julia_setup}.}
}
\description{
\code{dmm.setup} does the initial setup for jDirichletMixtureModels package.
}
\examples{
\dontrun{
DMM.setup()
}

}
