% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmm_cluster.R
\name{dmm.states}
\alias{dmm.states}
\title{Constructing list of all states from dmm.cluster run (excluding burnin).}
\usage{
dmm.states(juliastates, paramnames = NULL)
}
\arguments{
\item{juliastates}{The object returned by Julia code}

\item{paramnames}{Optionally. A list of the parameter names. Returned by Julia code for most bulit-in models.}
}
\description{
Constructing list of all states from dmm.cluster run (excluding burnin).
}
\details{
Each item in the list (i.e. \code{state = states[[i]]}) is a state. A state is also a list.
A state has two fields: \code{labeledData} and \code{clusterInfo}.

\code{labeledData} is a data.frame of the data points and their cluster labels.
\code{clusterInfo} is either a list or a data.table (if the data.table package is loaded by the user). It conatins
(1) cluster labels, (2) the number of data points (i.e. population) of each cluster, and (3) all of the parameters for each cluster.

If clusterInfo is a data.table, each row refers to a cluster. Columns are the cluster label, the population, and the rest of the columns are parameters.

If clusterInfo is a list, each element of the list refers to a clsuter,  clusterInfo[[i]] is a list containing of the above information for
cluster i as elements. E.g. clusterInfo[[1]]$population is the population of cluster 1. The params field (clusterInfo[[i]]$params)
is itself a list of each of the parameters
}
